Reorganize model files
