+++ graphient/src/main/scala/graphient/GraphqlClientError.scala
@@ -1,3 +1,3 @@
 package graphient
 
-case class GraphqlClientError(message: String) extends Exception(message)
+class GraphqlClientError {}
