--- graphient/src/main/scala/graphient/GraphqlRequest.scala
@@ -1,3 +1,0 @@
-package graphient
-
-case class GraphqlRequest[T](query: String, variables: T)
