--- graphient/src/main/scala/graphient/GraphqlResponseError.scala
@@ -1,4 +1,0 @@
-package graphient
-
-case class GraphqlResponseError(message: String, path: List[String], locations: List[GraphqlResponseErrorLocation])
-    extends Throwable
