--- graphient/src/main/scala/graphient/GraphqlResponseErrorLocation.scala
@@ -1,3 +1,0 @@
-package graphient
-
-case class GraphqlResponseErrorLocation(line: Int, column: Int)
