--- graphient/src/main/scala/graphient/RawGraphqlResponse.scala
@@ -1,6 +1,0 @@
-package graphient
-
-case class RawGraphqlResponse[T](
-    data:   Option[Map[String, T]],
-    errors: Option[List[GraphqlResponseError]]
-)
