+++ graphient/src/test/scala/graphient/specs/VariableGeneratorSpec.scala
@@ -1,6 +1,7 @@
 package graphient.specs
 
 import graphient._
+import graphient.model._
 import graphient.TestSchema.Domain._
 import org.scalatest._
 import sangria.ast
